+++ decidim-surveys/spec/system/private_space_survey_spec.rb
@@ -32,25 +32,11 @@ describe "Private Space Answer a survey", type: :system do
 
   let!(:participatory_space) { participatory_space_private }
 
-  let!(:component) do
-    create(:component,
-           manifest: manifest,
-           participatory_space: participatory_space)
-  end
+  let!(:component) { create(:component, manifest: manifest, participatory_space: participatory_space) }
 
   before do
     switch_to_host(organization.host)
-    puts "-------------"
-    puts "#{component.settings.attributes}"
-    puts "#{component.step.attributes}"
-    puts "-------------"
-    # component.update!(
-    #   step_settings: {
-    #     component.participatory_space.active_step.id => {
-    #       allow_answers: true
-    #     }
-    #   }
-    # )
+    component.update!(default_step_settings: { allow_answers: true })
   end
 
   def visit_component
@@ -59,6 +45,7 @@ describe "Private Space Answer a survey", type: :system do
 
   context "when space is private and transparent" do
     let!(:participatory_space_private) { create :assembly, :published, organization: organization, private_space: true, is_transparent: true }
+
     context "when the user is not logged in" do
       it "does not allow answering the survey" do
         visit_component
@@ -73,6 +60,7 @@ describe "Private Space Answer a survey", type: :system do
         end
       end
     end
+
     context "when the user is logged in" do
       context "and is private user space" do
         before do
@@ -97,7 +85,6 @@ describe "Private Space Answer a survey", type: :system do
           expect(page).to have_content("You have already answered this survey.")
           expect(page).to have_no_i18n_content(survey_question.body)
         end
-
       end
 
       context "and is not private user space" do
@@ -114,8 +101,6 @@ describe "Private Space Answer a survey", type: :system do
             expect(page).to have_content "The survey is available only for private users"
             expect(page).to have_content "Survey closed"
 
-            expect(page).to have_selector("survey[answers][0][body][disabled]")
-
             expect(page).to have_selector(".button[disabled]")
           end
         end
@@ -125,5 +110,49 @@ describe "Private Space Answer a survey", type: :system do
 
   context "when the spaces is private and not transparent" do
     let!(:participatory_space_private) { create :assembly, :published, organization: organization, private_space: true, is_transparent: false }
+
+    context "when the user is not logged in" do
+      it_behaves_like "a 404 page" do
+        let(:target_path) { main_component_path(component) }
+      end
+    end
+
+    context "when the user is logged in" do
+      context "and is private user space" do
+        before do
+          login_as other_user, scope: :user
+        end
+
+        it "allows answering the survey" do
+          visit_component
+
+          expect(page).to have_i18n_content(survey.title, upcase: true)
+          expect(page).to have_i18n_content(survey.description)
+
+          fill_in survey_question.body["en"], with: "My first answer"
+
+          check "survey_tos_agreement"
+
+          accept_confirm { click_button "Submit" }
+
+          within ".success.flash" do
+            expect(page).to have_content("successfully")
+          end
+
+          expect(page).to have_content("You have already answered this survey.")
+          expect(page).to have_no_i18n_content(survey_question.body)
+        end
+      end
+
+      context "and is not private user space" do
+        before do
+          login_as user, scope: :user
+        end
+
+        it_behaves_like "a 404 page" do
+          let(:target_path) { main_component_path(component) }
+        end
+      end
+    end
   end
 end